Split ImageFormatService into multiple files
