+++ OpenKh.Tools.ImageViewer/Services/ImageFormatService.cs
@@ -7,93 +7,8 @@ using OpenKh.Kh2;
 
 namespace OpenKh.Tools.ImageViewer.Services
 {
-    public class ImageFormatService : IImageFormatService
+    public partial class ImageFormatService : IImageFormatService
     {
-        private class GenericImageFormat : IImageFormat
-        {
-            private readonly Func<Stream, bool> isValid;
-
-            public GenericImageFormat(
-                string name,
-                string ext,
-                bool isContainer,
-                bool isCreationSupported,
-                Func<Stream, bool> isValid)
-            {
-                Name = name;
-                Extension = ext;
-                IsContainer = isContainer;
-                IsCreationSupported = isCreationSupported;
-                this.isValid = isValid;
-            }
-
-            public string Name { get; }
-            public string Extension { get; }
-            public bool IsContainer { get; }
-            public bool IsCreationSupported { get; }
-            public T As<T>() where T : IImageFormat => (T)(object)this;
-
-            public bool IsValid(Stream stream) => isValid(stream);
-        }
-
-        private class SingleImageFormat : GenericImageFormat, IImageSingle
-        {
-            private readonly Func<Stream, IImageRead> read;
-            private readonly Action<Stream, IImageRead> write;
-
-            public SingleImageFormat(
-                string name,
-                string ext,
-                bool isCreationSupported,
-                Func<Stream, bool> isValid,
-                Func<Stream, IImageRead> read,
-                Action<Stream, IImageRead> write) :
-                base(name, ext, false, isCreationSupported, isValid)
-            {
-                this.read = read;
-                this.write = write;
-            }
-
-            public IImageRead Read(Stream stream) => read(stream);
-            public void Write(Stream stream, IImageRead image) => write(stream, image);
-        }
-
-        private class MultipleImageFormat : GenericImageFormat, IImageMultiple
-        {
-            private readonly Func<Stream, IImageContainer> read;
-            private readonly Action<Stream, IImageContainer> write;
-
-            public MultipleImageFormat(
-                string name,
-                string ext,
-                bool isCreationSupported,
-                Func<Stream, bool> isValid,
-                Func<Stream, IImageContainer> read,
-                Action<Stream, IImageContainer> write) :
-                base(name, ext, isCreationSupported, true, isValid)
-            {
-                this.read = read;
-                this.write = write;
-            }
-
-            public IImageContainer Read(Stream stream) => read(stream);
-            public void Write(Stream stream, IImageContainer image) => write(stream, image);
-        }
-
-        private class ImageContainer : IImageContainer
-        {
-            private readonly IImageRead[] _images;
-
-            public ImageContainer(IEnumerable<IImageRead> images)
-            {
-                _images = images.ToArray();
-            }
-
-            public int Count => _images.Length;
-            public IEnumerable<IImageRead> Images => _images;
-            public IImageRead GetImage(int index) => _images[index];
-        }
-
         private static readonly IImageFormat[] imageFormat;
 
         static ImageFormatService()
